backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
requests
pymongo
python-dotenv
